+++ core/build.gradle.kts
@@ -117,6 +117,7 @@ kotlin {
         commonTest.dependencies {
             implementation(libs.kotlin.test)
             implementation(libs.test.coroutines)
+            implementation(libs.kermit.test)
         }
     }
 }
