+++ core/src/commonTest/kotlin/com/powersync/sync/SyncStreamTest.kt
@@ -1,0 +1,92 @@
+package com.powersync.sync
+
+import kotlin.test.*
+import kotlinx.serialization.json.JsonObject
+import co.touchlab.kermit.Logger
+import co.touchlab.kermit.Severity
+import co.touchlab.kermit.TestConfig
+import co.touchlab.kermit.TestLogWriter
+import com.powersync.bucket.BucketStorage
+import com.powersync.connectors.PowerSyncBackendConnector
+import com.powersync.db.crud.CrudEntry
+import com.powersync.db.crud.UpdateType
+import dev.mokkery.answering.returns
+import dev.mokkery.everySuspend
+import dev.mokkery.mock
+import dev.mokkery.resetAnswers
+import dev.mokkery.verify
+import kotlinx.coroutines.test.runTest
+
+class SyncStreamTest {
+    private lateinit var bucketStorage: BucketStorage
+    private lateinit var connector: PowerSyncBackendConnector
+    private lateinit var syncStream: SyncStream
+    private val testLogWriter = TestLogWriter(
+        loggable = Severity.Verbose
+    )
+    private val logger = Logger(
+        TestConfig(
+            minSeverity = Severity.Debug,
+            logWriterList = listOf(testLogWriter)
+        )
+    )
+
+    @BeforeTest
+    fun setup() {
+        bucketStorage = mock<BucketStorage>()
+        connector = mock<PowerSyncBackendConnector>()
+    }
+
+    @Test
+    fun testInvalidateCredentials() = runTest {
+        connector = mock<PowerSyncBackendConnector>() {
+            everySuspend { invalidateCredentials() } returns  Unit
+        }
+
+        syncStream = SyncStream(
+            bucketStorage = bucketStorage,
+            connector = connector,
+            uploadCrud = {},
+            logger = logger,
+            params = JsonObject(emptyMap())
+        )
+
+        syncStream.invalidateCredentials()
+        verify { connector.invalidateCredentials() }
+    }
+
+    // TODO: Work on improving testing this without needing to test the logs are displayed
+    @Test
+    fun testTriggerCrudUploadWhenAlreadyUploading() = runTest {
+        val mockCrudEntry = CrudEntry(id = "1", clientId = 1, op = UpdateType.PUT, table = "table1", transactionId = 1, opData = mapOf("key" to "value"))
+        bucketStorage = mock<BucketStorage>() {
+            everySuspend { nextCrudItem() } returns mockCrudEntry
+        }
+
+        syncStream = SyncStream(
+            bucketStorage = bucketStorage,
+            connector = connector,
+            uploadCrud = { },
+            retryDelayMs = 10,
+            logger = logger,
+            params = JsonObject(emptyMap())
+        )
+
+        syncStream.status.update(connected = true)
+        syncStream.triggerCrudUpload()
+
+        testLogWriter.assertCount(2)
+
+        with(testLogWriter.logs[0]) {
+            assertContains(
+                message,
+                "Potentially previously uploaded CRUD entries are still present in the upload queue."
+            )
+            assertEquals(Severity.Warn, severity)
+        }
+        with(testLogWriter.logs[1]) {
+            assertEquals(message,"Error uploading crud")
+            assertEquals(Severity.Error, severity)
+        }
+    }
+}
